requirements.txt
beautifulsoup4==4.9.0
feedparser==5.2.1
python-telegram-bot==12.7